requirements.txt
accelerate==0.31.0
bitsandbytes==0.44.1
blobfile==2.1.1
clean_fid==0.1.35
clip==1.0
einops==0.8.0
huggingface_hub==0.23.4
lmdb==1.4.1
mpi4py==4.0.1
natsort==8.4.0
numpy==1.24.3
openai_clip==1.0.1
Pillow==11.0.0
PyYAML==6.0.1
safetensors==0.4.5
simple_parsing==0.1.5
tensorflow==2.18.0
torch==2.3.1+cu118
torchvision==0.18.1+cu118
tqdm==4.66.5
transformers==4.37.2
triton==2.3.1
